code improvement for water cell
