+++ src/main/java/com/glodblock/github/FluidCraft.java
@@ -78,7 +78,7 @@ public class FluidCraft {
         }
         if (!ModAndClassUtil.NEW_HORIZONS_CORE_MOD) {
             RecipeLoader.addTerminalCards();
-            RecipeLoader.addRecipesIfDreamCoreNotExist();
+            RecipeLoader.addInfiniteWaterCell();
         }
         RecipeLoader.runTerminalRecipe();
