+++ src/main/java/com/glodblock/github/common/item/ItemCreativeFluidStorageCell.java
@@ -6,14 +6,17 @@ import net.minecraft.inventory.IInventory;
 import net.minecraft.item.ItemStack;
 
 import com.glodblock.github.FluidCraft;
+import com.glodblock.github.api.FluidCraftAPI;
 import com.glodblock.github.common.storage.IStorageFluidCell;
 import com.glodblock.github.common.tabs.FluidCraftingTabs;
 import com.glodblock.github.loader.IRegister;
 import com.glodblock.github.util.NameConst;
 
 import appeng.api.config.FuzzyMode;
+import appeng.api.storage.data.IAEFluidStack;
 import appeng.core.features.AEFeature;
 import appeng.items.contents.CellConfig;
+import appeng.items.contents.CellUpgrades;
 import appeng.util.Platform;
 import cpw.mods.fml.common.registry.GameRegistry;
 
@@ -59,4 +62,46 @@ public class ItemCreativeFluidStorageCell extends ItemBaseInfinityStorageCell
         setCreativeTab(FluidCraftingTabs.INSTANCE);
         return this;
     }
+
+    @Override
+    public long getBytes(ItemStack cellItem) {
+        return 0;
+    }
+
+    @Override
+    public int getBytesPerType(ItemStack cellItem) {
+        return 0;
+    }
+
+    @Override
+    public boolean isBlackListed(ItemStack cellItem, IAEFluidStack requestedAddition) {
+        return requestedAddition == null || requestedAddition.getFluid() == null
+                || FluidCraftAPI.instance().isBlacklistedInStorage(requestedAddition.getFluid().getClass());
+    }
+
+    @Override
+    public boolean storableInStorageCell() {
+        return false;
+    }
+
+    @Override
+    public boolean isStorageCell(ItemStack i) {
+        return true;
+    }
+
+    @Override
+    public double getIdleDrain(ItemStack is) {
+        return 0;
+    }
+
+    @Override
+    public int getTotalTypes(ItemStack cellItem) {
+        return 0;
+    }
+
+    @Override
+    public IInventory getUpgradesInventory(ItemStack is) {
+        return new CellUpgrades(is, 0);
+    }
+
 }
