+++ src/main/java/com/glodblock/github/common/storage/CreativeFluidCellInventory.java
@@ -61,7 +61,7 @@ public class CreativeFluidCellInventory extends FluidCellInventory {
                 if (fs == null) continue;
                 IAEFluidStack iaeFluidStack = Util.FluidUtil.createAEFluidStack(fs);
                 if (this.cellFluids.findPrecise(iaeFluidStack) == null) {
-                    iaeFluidStack.setStackSize(Long.MAX_VALUE);
+                    iaeFluidStack.setStackSize(Integer.MAX_VALUE);
                     this.cellFluids.add(iaeFluidStack);
                 }
             }
